Fix weird `directory ... not found`
It turns out some OSs have `sh` linked to other kinds of shells, that are not
`bash`. For example, Ubuntu uses `dash` (ref:
https://wiki.ubuntu.com/DashAsBinSh).

Some of those shells do not provide the built-in command `source`, which causes
the function `JumpCommand` to fail with an 127 error code.

Since `jump.vim` assumes all errors are caused by non-existing directories,
the resulting error message was very enigmatic and frustrating…